+++ doc/whats_new.rst
@@ -39,6 +39,13 @@ Classifiers and regressors
 Bug fixes
 .........
 
+Preprocessing
+
+- Fix a bug to ensure that ``func`` and ``inverse_func`` return a 2-dimensional
+  array when ``validate=True`` in :class:`FunctionTransformer` to agree with
+  the scikit-learn estimator API. :issue:`9595`. By :user:`Guillaume Lemaitre
+  <glemaitre>`.
+
 Decomposition, manifold learning and clustering
 
 - Fix for uninformative error in :class:`decomposition.incremental_pca`: